Better Image Manipulation & Display
This corresponds to Objective 2 of the Fall development plan on page 7 of my lab notebook.

In order to make typing plantID's less cumbersome, The hotkeys to control image rotation, call the viewfinder, and reset the tagging now require use of the Control key.

The data labels and other on-screen indicators will be redrawn over the image after every manipulation so that they remain legible on top of the image.